f1-telemetry-producer/requirements.txt
fastf1
kafka-python
pandas
